Rename {advanced->metadata} search; add metadata search to CLI
