+++ cmd/tools/crossref-checker/main.go
@@ -219,7 +219,7 @@ func main() {
 	}
 	searchForThinHashesAsBytes := specifiedThinHashesAsBytes
 
-	files, err := a.MatchingFiles(inclGlob, "", nil, nil)
+	files, err := a.MatchingFiles(inclGlob, "", nil, nil, "")
 	if err != nil {
 		prt.Fatalf("Error matching files: %v", err)
 	}
