+++ cmd/tools/hd2grep/main.go
@@ -46,7 +46,7 @@ func main() {
 		prt.Fatalf("Error opening game dir: %v", err)
 	}
 
-	files, err := a.MatchingFiles(inclGlob, "", nil, nil)
+	files, err := a.MatchingFiles(inclGlob, "", nil, nil, "")
 	if err != nil {
 		prt.Fatalf("Error matching files: %v", err)
 	}
